# client/resources/.swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/centre": {
      "post": {
        "operationId": "CentreController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCentreDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CentreDto" }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          }
        },
        "tags": ["Centre"]
      },
      "get": {
        "operationId": "CentreController_getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/CentreDto" }
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          }
        },
        "tags": ["Centre"]
      }
    },
    "/api/centre/{centreId}/add-admin": {
      "post": {
        "operationId": "CentreController_addAdmin",
        "parameters": [
          {
            "name": "centreId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessDto" }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          }
        },
        "tags": ["Centre"]
      }
    },
    "/api/centre/all": {
      "get": {
        "operationId": "CentreController_getCentres",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/CentreDto" }
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          }
        },
        "tags": ["Centre"]
      }
    },
    "/api/centre/{id}": {
      "get": {
        "operationId": "CentreController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CentreDto" }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          }
        },
        "tags": ["Centre"]
      }
    },
    "/api/centre/expense": {
      "post": {
        "operationId": "ExpenseController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateExpenseDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ExpenseDto" }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          }
        },
        "tags": ["Centre Expense"]
      }
    },
    "/api/centre/expense/{centreId}": {
      "get": {
        "operationId": "ExpenseController_getAll",
        "parameters": [
          {
            "name": "centreId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ExpenseDto" }
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          }
        },
        "tags": ["Centre Expense"]
      }
    },
    "/api/centre/expense/{centreId}/{id}": {
      "get": {
        "operationId": "ExpenseController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "centreId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ExpenseDto" }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          }
        },
        "tags": ["Centre Expense"]
      }
    },
...
    "/api/centre/rate-list/{centreId}/centre": {
      "get": {
        "operationId": "RateListController_get",
        "parameters": [
          {
            "name": "centreId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/RateListDto" }
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          }
        },
        "tags": ["Centre Rate List"]
      }
    },
...
    "/api/doctor-commission/centre/{centreId}/all": {
      "get": {
        "operationId": "DoctorCommissionController_getAllDoctorsForCentre",
        "parameters": [
          {
            "name": "centreId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DoctorCommissionDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          }
        },
        "tags": ["DoctorCommission"]
      }
    },
...
    "/api/booking": {
      "post": {
        "operationId": "BookingController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateBookingDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BookingDto" }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          }
        },
        "tags": ["Booking"]
      }
    },
    "/api/booking/{id}": {
      "get": {
        "operationId": "BookingController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BookingDto" }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          }
        },
        "tags": ["Booking"]
      }
    },
    "/api/booking/centre/{centreId}": {
      "get": {
        "operationId": "BookingController_get",
        "parameters": [
          {
            "name": "centreId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/BookingDto" }
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          }
        },
        "tags": ["Booking"]
      }
    },
...
    "/api/update-request/{centreId}/centre": {
      "get": {
        "operationId": "UpdateRequestController_get",
        "parameters": [
          {
            "name": "centreId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/UpdateRequestDto" }
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          }
        },
        "tags": ["Update Request"]
      }
    }
  },
  "info": {
    "title": "Radiology API",
    "description": "Radiology booking API",
    "version": "v0.0.0",
    "contact": {
      "name": "Radiology",
      "url": "https://nadaaa.vercel.app/",
      "email": "nadaaofficial19@gmail.com"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
      "UserRole": {
        "type": "string",
        "enum": ["admin", "receptionist", "pr", "doctor"]
      },
...
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "c3611c05-df51-4b47-b601-f2eac02f4ef0"
          },
          "name": { "type": "string" },
          "email": { "type": "string" },
          "password": { "type": "string", "format": "^[\\d!#$%&*@A-Z^a-z]*$" },
          "centreId": {
            "type": "string",
            "format": "uuid",
            "example": "c3611c05-df51-4b47-b601-f2eac02f4ef0"
          },
          "role": { "$ref": "#/components/schemas/UserRole" }
        },
        "required": ["id", "name", "email", "password", "role"]
      },
...
      "ErrorDto": {
        "type": "object",
        "properties": {
          "message": { "type": "object" },
          "statusCode": { "type": "number" },
          "error": { "type": "string" }
        },
        "required": ["message", "statusCode", "error"]
      },
...
      "SuccessDto": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "statusCode": { "type": "number" }
        },
        "required": ["message", "statusCode"]
      },
      "AddressDto": {
        "type": "object",
        "properties": {
          "line1": { "type": "string" },
          "line2": { "type": "string" },
          "city": { "type": "string" },
          "postalCode": { "type": "number", "example": 1 },
          "state": { "type": "string" },
          "country": { "type": "string" }
        },
        "required": ["line1", "city", "postalCode", "state"]
      },
      "CreateCentreDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string" },
          "phone": { "type": "string" },
          "address": { "$ref": "#/components/schemas/AddressDto" }
        },
        "required": ["name", "email", "phone", "address"]
      },
      "CentreDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "c3611c05-df51-4b47-b601-f2eac02f4ef0"
          },
          "createdAt": { "format": "date-time", "type": "string" },
          "name": { "type": "string" },
          "email": { "type": "string" },
          "phone": { "type": "string" },
          "address": { "$ref": "#/components/schemas/AddressDto" }
        },
        "required": ["id", "createdAt", "name", "email", "phone", "address"]
      },
      "CreateExpenseDto": {
        "type": "object",
        "properties": {
          "centreId": {
            "type": "string",
            "format": "uuid",
            "example": "c3611c05-df51-4b47-b601-f2eac02f4ef0"
          },
          "date": { "format": "date-time", "type": "string" },
          "amount": { "type": "number", "example": 1 },
          "expenseType": { "type": "string" },
          "paymentMethod": { "type": "string" },
          "remark": { "type": "string" }
        },
        "required": [
          "centreId",
          "date",
          "amount",
          "expenseType",
          "paymentMethod"
        ]
      },
      "ExpenseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "c3611c05-df51-4b47-b601-f2eac02f4ef0"
          },
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "centreId": {
            "type": "string",
            "format": "uuid",
            "example": "c3611c05-df51-4b47-b601-f2eac02f4ef0"
          },
          "date": { "format": "date-time", "type": "string" },
          "createdBy": {
            "type": "string",
            "format": "uuid",
            "example": "c3611c05-df51-4b47-b601-f2eac02f4ef0"
          },
          "amount": { "type": "number", "example": 1 },
          "expenseType": { "type": "string" },
          "paymentMethod": { "type": "string" },
          "remark": { "type": "string" }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "centreId",
          "date",
          "createdBy",
          "amount",
          "expenseType",
          "paymentMethod"
        ]
      },
      "InvestigationDto": {
        "type": "object",
        "properties": {
          "type": { "type": "string" },
          "amount": { "type": "number", "example": 1 },
          "filmCount": { "type": "number", "example": 1 }
        },
        "required": ["type", "amount", "filmCount"]
      },
      "RateListDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "c3611c05-df51-4b47-b601-f2eac02f4ef0"
          },
          "centreId": {
            "type": "string",
            "format": "uuid",
            "example": "c3611c05-df51-4b47-b601-f2eac02f4ef0"
          },
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "modality": { "type": "string" },
          "investigation": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/InvestigationDto" }
          }
        },
        "required": [
          "id",
          "centreId",
          "createdAt",
          "updatedAt",
          "modality",
          "investigation"
        ]
      },
...
      "DoctorCommissionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "c3611c05-df51-4b47-b601-f2eac02f4ef0"
          },
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "doctorId": {
            "type": "string",
            "format": "uuid",
            "example": "c3611c05-df51-4b47-b601-f2eac02f4ef0"
          },
          "centreId": {
            "type": "string",
            "format": "uuid",
            "example": "c3611c05-df51-4b47-b601-f2eac02f4ef0"
          },
          "modality": { "type": "string" },
          "amount": { "type": "number", "example": 1 },
          "startDate": { "format": "date-time", "type": "string" },
          "endDate": { "format": "date-time", "type": "string" },
          "doctor": { "$ref": "#/components/schemas/UserDto" },
          "centre": { "$ref": "#/components/schemas/CentreDto" }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "doctorId",
          "centreId",
          "modality",
          "amount"
        ]
      },
...
      "CreatePatientDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "age": { "type": "number", "example": 1 },
          "gender": { "type": "string" },
          "phone": { "type": "string" },
          "email": { "type": "string" },
          "address": { "type": "string" },
          "abhaId": { "type": "string" }
        },
        "required": ["name", "age", "gender", "phone"]
      },
      "CreateBookingDto": {
        "type": "object",
        "properties": {
          "centreId": {
            "type": "string",
            "format": "uuid",
            "example": "c3611c05-df51-4b47-b601-f2eac02f4ef0"
          },
          "consultant": {
            "type": "string",
            "format": "uuid",
            "example": "c3611c05-df51-4b47-b601-f2eac02f4ef0"
          },
          "modality": { "type": "string" },
          "investigation": { "type": "string" },
          "amount": { "type": "number", "example": 1 },
          "discount": { "type": "number", "example": 1 },
          "remark": { "type": "string" },
          "extraCharge": { "type": "string" },
          "paymentType": { "type": "string" },
          "patientId": {
            "type": "string",
            "format": "uuid",
            "example": "c3611c05-df51-4b47-b601-f2eac02f4ef0"
          },
          "patient": { "$ref": "#/components/schemas/CreatePatientDto" }
        },
        "required": [
          "centreId",
          "consultant",
          "modality",
          "investigation",
          "amount",
          "paymentType"
        ]
      },
      "PatientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "c3611c05-df51-4b47-b601-f2eac02f4ef0"
          },
          "createdAt": { "format": "date-time", "type": "string" },
          "name": { "type": "string" },
          "age": { "type": "number", "example": 1 },
          "gender": { "type": "string" },
          "phone": { "type": "string" },
          "email": { "type": "string" },
          "address": { "type": "string" },
          "abhaId": { "type": "string" },
          "booking": { "type": "array", "items": { "type": "array" } }
        },
        "required": ["id", "createdAt", "name", "age", "gender", "phone"]
      },
      "BookingRecordDto": {
        "type": "object",
        "properties": {
          "type": { "type": "string" },
          "url": { "type": "string" }
        },
        "required": ["type", "url"]
      },
      "BookingDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "c3611c05-df51-4b47-b601-f2eac02f4ef0"
          },
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "centreId": {
            "type": "string",
            "format": "uuid",
            "example": "c3611c05-df51-4b47-b601-f2eac02f4ef0"
          },
          "patientId": {
            "type": "string",
            "format": "uuid",
            "example": "c3611c05-df51-4b47-b601-f2eac02f4ef0"
          },
          "submittedBy": {
            "type": "string",
            "format": "uuid",
            "example": "c3611c05-df51-4b47-b601-f2eac02f4ef0"
          },
          "consultant": {
            "type": "string",
            "format": "uuid",
            "example": "c3611c05-df51-4b47-b601-f2eac02f4ef0"
          },
          "modality": { "type": "string" },
          "investigation": { "type": "string" },
          "amount": { "type": "number", "example": 1 },
          "discount": { "type": "number", "example": 1 },
          "remark": { "type": "string" },
          "extraCharge": { "type": "string" },
          "paymentType": { "type": "string" },
          "patient": { "$ref": "#/components/schemas/PatientDto" },
          "record": { "$ref": "#/components/schemas/BookingRecordDto" }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "centreId",
          "patientId",
          "submittedBy",
          "consultant",
          "modality",
          "investigation",
          "amount",
          "paymentType"
        ]
      },
      "RequestType": { "type": "string", "enum": ["expense", "booking"] },
      "RequestStatus": {
        "type": "string",
        "enum": ["pending", "accepted", "rejected"]
      },
      "UpdateBookingDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "c3611c05-df51-4b47-b601-f2eac02f4ef0"
          },
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "centreId": {
            "type": "string",
            "format": "uuid",
            "example": "c3611c05-df51-4b47-b601-f2eac02f4ef0"
          },
          "patientId": {
            "type": "string",
            "format": "uuid",
            "example": "c3611c05-df51-4b47-b601-f2eac02f4ef0"
          },
          "submittedBy": {
            "type": "string",
            "format": "uuid",
            "example": "c3611c05-df51-4b47-b601-f2eac02f4ef0"
          },
          "consultant": {
            "type": "string",
            "format": "uuid",
            "example": "c3611c05-df51-4b47-b601-f2eac02f4ef0"
          },
          "modality": { "type": "string" },
          "investigation": { "type": "string" },
          "amount": { "type": "number", "example": 1 },
          "discount": { "type": "number", "example": 1 },
          "remark": { "type": "string" },
          "extraCharge": { "type": "string" },
          "paymentType": { "type": "string" },
          "record": { "$ref": "#/components/schemas/BookingRecordDto" }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "centreId",
          "patientId",
          "submittedBy",
          "consultant",
          "modality",
          "investigation",
          "amount",
          "paymentType"
        ]
      },
...
      "UpdateRequestDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "c3611c05-df51-4b47-b601-f2eac02f4ef0"
          },
          "centreId": {
            "type": "string",
            "format": "uuid",
            "example": "c3611c05-df51-4b47-b601-f2eac02f4ef0"
          },
          "createdAt": { "format": "date-time", "type": "string" },
          "requestedBy": {
            "type": "string",
            "format": "uuid",
            "example": "c3611c05-df51-4b47-b601-f2eac02f4ef0"
          },
          "type": { "$ref": "#/components/schemas/RequestType" },
          "status": { "$ref": "#/components/schemas/RequestStatus" },
          "expenseData": { "$ref": "#/components/schemas/ExpenseDto" },
          "bookingData": { "$ref": "#/components/schemas/UpdateBookingDto" }
        },
        "required": [
          "id",
          "centreId",
          "createdAt",
          "requestedBy",
          "type",
          "status"
        ]
      }
    }
  }
}